Add new group container + moving existind AddLabels transformation to this group
